自定义session添加部门属性

+++ crowdfounding-framework/src/main/java/com/wayn/framework/shiro/realm/MyRealm.java
@@ -2,6 +2,8 @@ package com.wayn.framework.shiro.realm;
 
 import com.baomidou.mybatisplus.mapper.EntityWrapper;
 import com.wayn.commom.domain.User;
+import com.wayn.commom.enums.StateEnum;
+import com.wayn.commom.exception.BusinessException;
 import com.wayn.commom.service.RoleMenuService;
 import com.wayn.commom.service.UserRoleService;
 import com.wayn.commom.service.UserService;
@@ -30,6 +32,9 @@ public class MyRealm extends AuthorizingRealm {
 
     @Value("${wayn.singeUserAuth}")
     private boolean singeUserAuth;
+
+    @Value("${wayn.singeKickoutBefore}")
+    private boolean singeKickoutBefore;
 
     /**
      * 用户服务
@@ -93,10 +98,11 @@ public class MyRealm extends AuthorizingRealm {
         if (sysUser.getUserState() == -1) {
             throw new UnknownAccountException("用户已被禁用");
         }
+        /// 此处不再需要做密码验证，由CredentialsMatch的实现做密码校验
         /*if (!sysUser.getPassword().equals(password)) {
             throw new IncorrectCredentialsException("账号或密码不正确");
         }*/
-        if (sysUser.getUserState() == User._0) {
+        if (sysUser.getUserState() == StateEnum.DISABLE.getState()) {
             throw new LockedAccountException("该用户已被锁定，请稍后再试");
         }
         // 是否进行单一用户登陆处理
@@ -110,12 +116,16 @@ public class MyRealm extends AuthorizingRealm {
                             .getAttribute(DefaultSubjectContext.PRINCIPALS_SESSION_KEY);
                     Object primaryPrincipal = principalCollection.getPrimaryPrincipal();
                     User user = (User) primaryPrincipal;
-                    if (user.getUserName().equals(username)) {
-                        Session session = sessionDAO.readSession(activeSession.getId());
-                        if (session != null) {
-                            session.stop();
-                            sessionDAO.delete(session);
-                            simpMessagingTemplate.convertAndSendToUser(user.getId(), "/queue/getResponse", "新消息：" + "该账号已在其他机器登陆！");
+                    if (user.getUserName().equals(username) && singeKickoutBefore) {
+                        if (singeKickoutBefore) {
+                            Session session = sessionDAO.readSession(activeSession.getId());
+                            if (session != null) {
+                                session.stop();
+                                sessionDAO.delete(session);
+                                simpMessagingTemplate.convertAndSendToUser(user.getId(), "/queue/getResponse", "新消息：" + "该账号已在其他机器登陆！");
+                            }
+                        } else {
+                            throw new BusinessException("该用户已登陆，请先登出！");
                         }
                     }
                 }
